Creation of Splash Screen and Card Details screen and request logic
Fragment was created for Splash Screen startup screen.
Fragment was created for the cars details screen.
Data layer was created.
Network layer was created.
Domain layer was created.
Presentation layer has been created.
Dependency injection was created to provide modules for the application.
A Selead Class was created to represent the view states.
Navigation for application Fragments was created.